+++ src/Engines/EdaSimulator.Engines/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace EdaSimulator.Engines;
-
-public class Class1
-{
-
-}